Task 14

+++ 0x03-python-data_structures/100-print_python_list_info.c
@@ -11,7 +11,7 @@
 void print_python_list_info(PyObject *p)
 {
 	PyObject *item;
-	Py_ssize_t l_size, i;
+	long int l_size, i;
 
 	l_size = PyList_Size(p);
 	printf("[*] Size of the Python List = %ld\n", l_size);
